Moved functions for rendering spec to rst to a separate python module. Renamed the rst module
